Add initial version of minimial opener software
This is based on secplus_reader.py, but strips away all the QT Gui and
just operates as a local CLI application, printing the detected results
to the console.

There is some minor re-organization of the Connections to make it more
readable, and hard-coding of the frequency appropriate for my use case.